chore: update CI workflows to use variable for ALLOW_DB_TRUNCATE
This commit modifies the CI workflow files for development and staging environments to replace the hardcoded value of `ALLOW_DB_TRUNCATE` with a reference to the variable `${{ vars.ALLOW_DB_TRUNCATE }}`. This change enhances the flexibility of the workflows by allowing the value to be set dynamically based on the environment, improving maintainability and consistency across different CI configurations.

Changes:
- Updated `ci-dev.yml` to use `${{ vars.ALLOW_DB_TRUNCATE }}` for `ALLOW_DB_TRUNCATE`
- Updated `ci-staging.yml` to use `${{ vars.ALLOW_DB_TRUNCATE }}` for `ALLOW_DB_TRUNCATE`